requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
idna==3.3
influxdb==5.3.1
msgpack==1.0.4
paho-mqtt==1.6.1
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
six==1.16.0
urllib3==1.26.11
